Move ingest task status store code to core module
