+++ java/clients/src/main/java/sleeper/clients/admin/AdminClientTrackerFactory.java
@@ -24,9 +24,9 @@ import sleeper.core.properties.table.TablePropertiesProvider;
 import sleeper.core.tracker.compaction.job.CompactionJobTracker;
 import sleeper.core.tracker.compaction.task.CompactionTaskTracker;
 import sleeper.core.tracker.ingest.job.IngestJobTracker;
+import sleeper.core.tracker.ingest.task.IngestTaskStatusStore;
 import sleeper.ingest.batcher.core.IngestBatcherStore;
 import sleeper.ingest.batcher.store.IngestBatcherStoreFactory;
-import sleeper.ingest.core.task.IngestTaskStatusStore;
 import sleeper.ingest.status.store.job.IngestJobTrackerFactory;
 import sleeper.ingest.status.store.task.IngestTaskStatusStoreFactory;
