+++ java/clients/src/main/java/sleeper/clients/status/report/IngestTaskStatusReport.java
@@ -25,7 +25,7 @@ import sleeper.clients.status.report.ingest.task.IngestTaskStatusReportArguments
 import sleeper.clients.status.report.ingest.task.IngestTaskStatusReporter;
 import sleeper.configuration.properties.S3InstanceProperties;
 import sleeper.core.properties.instance.InstanceProperties;
-import sleeper.ingest.core.task.IngestTaskStatusStore;
+import sleeper.core.tracker.ingest.task.IngestTaskStatusStore;
 import sleeper.ingest.status.store.task.IngestTaskStatusStoreFactory;
 
 import static sleeper.configuration.utils.AwsV1ClientHelper.buildAwsV1Client;
