+++ java/clients/src/main/java/sleeper/clients/status/report/ingest/task/IngestTaskStatusReporter.java
@@ -16,7 +16,7 @@
 
 package sleeper.clients.status.report.ingest.task;
 
-import sleeper.ingest.core.task.IngestTaskStatus;
+import sleeper.core.tracker.ingest.task.IngestTaskStatus;
 
 import java.io.PrintStream;
 import java.util.List;
