+++ java/core/src/main/java/sleeper/core/tracker/ingest/task/IngestTaskFinishedStatus.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.ingest.core.task;
+package sleeper.core.tracker.ingest.task;
 
 import sleeper.core.record.process.AverageRecordRate;
 import sleeper.core.record.process.RecordsProcessed;
