+++ java/core/src/main/java/sleeper/core/tracker/ingest/task/IngestTaskStatusesBuilder.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.ingest.core.task;
+package sleeper.core.tracker.ingest.task;
 
 import java.time.Instant;
 import java.util.Comparator;
