+++ java/core/src/test/java/sleeper/core/tracker/ingest/task/IngestTaskStatusInPeriodTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.ingest.core.task;
+package sleeper.core.tracker.ingest.task;
 
 import org.junit.jupiter.api.Test;
