+++ java/core/src/test/java/sleeper/core/tracker/ingest/task/IngestTaskStatusTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.ingest.core.task;
+package sleeper.core.tracker.ingest.task;
 
 import org.junit.jupiter.api.Test;
