+++ java/ingest/ingest-core/src/test/java/sleeper/ingest/core/IngestTaskTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.ingest.core.task;
+package sleeper.ingest.core;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -24,11 +24,12 @@ import sleeper.core.record.process.RecordsProcessedSummary;
 import sleeper.core.record.process.status.ProcessStatusUpdateRecord;
 import sleeper.core.tracker.ingest.job.InMemoryIngestJobTracker;
 import sleeper.core.tracker.ingest.job.IngestJobUpdateType;
-import sleeper.ingest.core.IngestResult;
+import sleeper.core.tracker.ingest.task.InMemoryIngestTaskStatusStore;
+import sleeper.core.tracker.ingest.task.IngestTaskStatusStore;
+import sleeper.ingest.core.IngestTask.MessageHandle;
+import sleeper.ingest.core.IngestTask.MessageReceiver;
 import sleeper.ingest.core.job.IngestJob;
 import sleeper.ingest.core.job.IngestJobHandler;
-import sleeper.ingest.core.task.IngestTask.MessageHandle;
-import sleeper.ingest.core.task.IngestTask.MessageReceiver;
 
 import java.time.Instant;
 import java.util.ArrayList;
@@ -46,14 +47,14 @@ import static org.assertj.core.api.Assertions.tuple;
 import static sleeper.core.tracker.ingest.job.IngestJobUpdateType.FAILED;
 import static sleeper.core.tracker.ingest.job.IngestJobUpdateType.FINISHED_WHEN_FILES_COMMITTED;
 import static sleeper.core.tracker.ingest.job.IngestJobUpdateType.STARTED;
+import static sleeper.core.tracker.ingest.task.IngestTaskStatusTestData.finishedMultipleJobs;
+import static sleeper.core.tracker.ingest.task.IngestTaskStatusTestData.finishedNoJobs;
+import static sleeper.core.tracker.ingest.task.IngestTaskStatusTestData.finishedOneJob;
 import static sleeper.ingest.core.IngestResultTestData.defaultFileIngestResult;
 import static sleeper.ingest.core.IngestResultTestData.defaultFileIngestResultReadAndWritten;
 import static sleeper.ingest.core.job.IngestJobStatusFromJobTestData.failedIngestJob;
 import static sleeper.ingest.core.job.IngestJobStatusFromJobTestData.finishedIngestJobUncommitted;
 import static sleeper.ingest.core.job.IngestJobTestData.DEFAULT_TABLE_ID;
-import static sleeper.ingest.core.task.IngestTaskStatusTestData.finishedMultipleJobs;
-import static sleeper.ingest.core.task.IngestTaskStatusTestData.finishedNoJobs;
-import static sleeper.ingest.core.task.IngestTaskStatusTestData.finishedOneJob;
 
 public class IngestTaskTest {
     private static final String DEFAULT_TASK_ID = "test-task-id";
