+++ java/ingest/ingest-runner/src/main/java/sleeper/ingest/runner/task/IngestJobQueueConsumer.java
@@ -32,10 +32,10 @@ import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.record.process.RecordsProcessedSummary;
 import sleeper.core.table.TableIndex;
 import sleeper.core.tracker.ingest.job.IngestJobTracker;
+import sleeper.ingest.core.IngestTask.MessageHandle;
+import sleeper.ingest.core.IngestTask.MessageReceiver;
 import sleeper.ingest.core.job.IngestJob;
 import sleeper.ingest.core.job.IngestJobMessageHandler;
-import sleeper.ingest.core.task.IngestTask.MessageHandle;
-import sleeper.ingest.core.task.IngestTask.MessageReceiver;
 import sleeper.job.common.action.ActionException;
 import sleeper.job.common.action.MessageReference;
 import sleeper.job.common.action.thread.PeriodicActionRunnable;
