+++ java/ingest/ingest-status-store/src/main/java/sleeper/ingest/status/store/task/IngestTaskStatusStoreFactory.java
@@ -19,7 +19,7 @@ package sleeper.ingest.status.store.task;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 
 import sleeper.core.properties.instance.InstanceProperties;
-import sleeper.ingest.core.task.IngestTaskStatusStore;
+import sleeper.core.tracker.ingest.task.IngestTaskStatusStore;
 
 import static sleeper.core.properties.instance.IngestProperty.INGEST_TRACKER_ENABLED;
