+++ java/ingest/ingest-status-store/src/test/java/sleeper/ingest/status/store/task/StoreIngestTaskIT.java
@@ -18,7 +18,7 @@ package sleeper.ingest.status.store.task;
 
 import org.junit.jupiter.api.Test;
 
-import sleeper.ingest.core.task.IngestTaskStatus;
+import sleeper.core.tracker.ingest.task.IngestTaskStatus;
 import sleeper.ingest.status.store.testutils.DynamoDBIngestTaskStatusStoreTestBase;
 
 import static org.assertj.core.api.Assertions.assertThat;
