+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/ingest/AwsIngestReportsDriver.java
@@ -28,7 +28,7 @@ import sleeper.clients.status.report.ingest.task.StandardIngestTaskStatusReporte
 import sleeper.clients.status.report.job.query.RangeJobsQuery;
 import sleeper.core.tracker.ingest.job.IngestJobStatus;
 import sleeper.core.tracker.ingest.job.IngestJobTracker;
-import sleeper.ingest.core.task.IngestTaskStatusStore;
+import sleeper.core.tracker.ingest.task.IngestTaskStatusStore;
 import sleeper.ingest.status.store.job.IngestJobTrackerFactory;
 import sleeper.ingest.status.store.task.IngestTaskStatusStoreFactory;
 import sleeper.systemtest.drivers.util.SystemTestClients;
