+++ java/system-test/system-test-dsl/src/test/java/sleeper/systemtest/dsl/testutil/drivers/InMemoryIngestByQueue.java
@@ -24,13 +24,13 @@ import sleeper.core.statestore.StateStore;
 import sleeper.core.table.TableIndex;
 import sleeper.core.tracker.ingest.job.InMemoryIngestJobTracker;
 import sleeper.core.tracker.ingest.job.IngestJobTracker;
+import sleeper.core.tracker.ingest.task.InMemoryIngestTaskStatusStore;
+import sleeper.core.tracker.ingest.task.IngestTaskStatusStore;
 import sleeper.ingest.core.IngestResult;
+import sleeper.ingest.core.IngestTask;
 import sleeper.ingest.core.job.IngestJob;
 import sleeper.ingest.core.job.IngestJobHandler;
 import sleeper.ingest.core.job.IngestJobMessageHandler;
-import sleeper.ingest.core.task.InMemoryIngestTaskStatusStore;
-import sleeper.ingest.core.task.IngestTask;
-import sleeper.ingest.core.task.IngestTaskStatusStore;
 import sleeper.ingest.runner.impl.commit.AddFilesToStateStore;
 import sleeper.query.core.recordretrieval.InMemoryDataStore;
 import sleeper.systemtest.dsl.SystemTestContext;
